[Breaking change] Follow k8s API conventions (#136)
This PR updates the Temporal Worker Controller APIs to follow Kubernetes
API conventions documented in
https://github.com/kubernetes/community/blob/master/contributors/devel/sig-architecture/api-conventions.md.

## What changed

- Added standard `conditions` field to both status structs per k8s API
conventions
- Refactored `RampPercentage` in rollout spec from float to int
- Changed connection references from plain strings to "ref" types
(`connectionRef`, `mutualTLSSecretRef`)
- Removed `Selector` field from `TemporalWorkerDeploymentSpec` (deployment
selectors are now computed by the controller directly, avoiding
possibility for misconfiguration)
- Added validation kubebuilder annotations to several fields
- Updated demo YAML examples to use new field names

## Breaking changes

This requires updating existing YAML files due to field name changes and
reinstalling the CRDs.

## Checklist

1. Closes #103 
2. How was this tested: unit/integration tests should pass
3. Any docs updates needed?

+++ internal/controller/execplan.go
@@ -105,7 +105,7 @@ func (r *TemporalWorkerDeploymentReconciler) executePlan(ctx context.Context, l 
 
 			if _, err := deploymentHandler.SetRampingVersion(ctx, sdkclient.WorkerDeploymentSetRampingVersionOptions{
 				BuildID:       vcfg.BuildID,
-				Percentage:    vcfg.RampPercentage,
+				Percentage:    float32(vcfg.RampPercentage),
 				ConflictToken: vcfg.ConflictToken,
 				Identity:      getControllerIdentity(),
 			}); err != nil {
